Update the structure
Change the structure of the project
