+++ src/main/java/springboot/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package springboot.util;
+package springboot.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
