+++ src/main/java/springboot/repository/IdealRepository.java
@@ -1,6 +1,8 @@
-package springboot.domain;
+package springboot.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import springboot.domain.Ideal;
 
 /**
  * <p>
