+++ src/main/java/springboot/repository/PlayerRepository.java
@@ -1,8 +1,10 @@
-package springboot.domain;
+package springboot.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import springboot.domain.Player;
 
 /**
  * <p>
