+++ src/main/java/springboot/repository/UserRepository.java
@@ -1,8 +1,10 @@
-package springboot.domain;
+package springboot.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import springboot.domain.User;
 
 /**
  * <p>
