+++ src/main/java/springboot/web/BattleController.java
@@ -14,7 +14,7 @@ import com.google.gson.JsonParser;
 
 import springboot.domain.Bot;
 import springboot.domain.Player;
-import springboot.service.BattleService;
+import springboot.service.impl.BattleService;
 
 @RestController
 public class BattleController {
